✨🗃Create StudentLogin databases
Create a class for student login data (StudentLogin) and another
'holding' class (StudentLoginTable) to store these objects.

The StudentLogin class has a __repr__ method for debugging and a
tabulate method to output data as a string for tabulation.
The StudentLoginTable class has a __repr__ method for debugging and
methods to save and load data from text files.