🗃️💥Reorganise class structure significantly
The Row and Table base classes now have much more purpose.
The RowClass has a key_field attribute to be overridden - this specifies
what attribute of the class should be unique.
The TableClass has a row_class attribute to be overridden - this
specifies what the class will be for objects to be stored in the table.

The methods for both base classes (tabulate(), load_from_file(), etc.)
have been greatly generalised and placed in the base class definition.

The tabulate method has been rewritten to dynamically fetch all an
object's attributes and then pad these into a string as specified by
padding_values.