♻️Improve datetime variable/str handling
Add new functions to calculate an end date and a DateInfoDict into a
readable string.
Date and string related functions are now much more versatile by being
able to take a separator arguments to customise look of dates (e.g.
YYYY/MM/DD vs YYYY-MM-DD).
The offset parameters validate_date() are now contained in a tuple
instead of as two separate arguments. This is to accommodate the new sep
argument.