✨Outline of staff Student Overview table GUI
Begin outline of Student Overview table for staff to manage students
enrolled in the award.
Create a function to add a hover-underline link to a label. On hover,
this label is underlined and when clicked redirects the user. This is
in anticipation of staff being able to click a student's name to view
more details.